fix: Use u128 for sizes (#52)

+++ src/state/file_to_delete.rs
@@ -9,7 +9,7 @@ pub struct FileToDelete {
     pub path_to_file: Vec<OsString>,
     pub file_type: FileType,
     pub num_descendants: Option<u64>,
-    pub size: u64,
+    pub size: u128,
 }
 
 impl FileToDelete {
